[tests] Get integration tests up and running.
It was necessary to switch from Jest to Mocha + Chai because Jest does not expose a stable API for programmatic invocation. Programmatic invocation is necessary in order for the tests to have access to the 'vscode' module, a process that involves actually running the tests within a VS Code instance. There is a backdoor way to invoke Jest programmatically, but I failed to find a way to do that while preserving access to the 'vscode' module.