Another ad hoc patch to type highlighting.
This is not really necessary because the programmatic/semantic
highlighter covers this for most color themes, but it does fix
highlighting for the "Light High Contrast" theme. This also matters for
the website and it reduces the delay before the correct highlighting is
applied.

This fix is "best effort." It will not cover all cases but overall it
will improve things in practice.